금욜 마무리  여행코스 통합

+++ src/main/java/project/dto/KoreapriceDto.java
@@ -7,5 +7,5 @@ public class KoreapriceDto {
 	//물가체험
 	int koreapriceIdx;			//한국물가상품인덱스
 	String koreapriceProduct;	//상품명
-	int koreapricePrice;		//상품가격
+	double koreapricePrice;		//상품가격
 }
